[form-factors] Move the transition tags to a separated header
